+++ client/src/App.js
@@ -6,11 +6,16 @@ import Submissions from './pages/Submissions'
 import Submission from './pages/Submission'
 import Vendors from './pages/Vendors'
 import SubmissionFormPage from './pages/SubmissionFormPage'
+import { ToastContainer } from 'react-toastify'
+import { useQuery } from '@tanstack/react-query'
+import 'react-toastify/dist/ReactToastify.css'
+import { getUserFromLocalStorage } from './helpers/localStorageHelper'
 
 import { Route, Routes } from 'react-router-dom'
 
-
 function App() {
+
+
   return (
     <div className="container m-auto">
       <Navbar />
@@ -23,6 +28,7 @@ function App() {
         <Route path="/vendors" element={<Vendors />} />
         <Route path="/:userId/form" element={<SubmissionFormPage />} />
       </Routes>
+      <ToastContainer />
     </div>
   )
 }