+++ client/src/Pages/Home.js
@@ -1,16 +1,12 @@
 import Hero from '../components/Hero'
 import axios from 'axios'
 import { useEffect } from 'react'
+import { ToastContainer } from 'react-toastify'
+import Alert from '../components/Alert'
+import { useSelector } from 'react-redux'
 
 const Home = () => {
-  const fetchData = async () => {
-    const response = await axios.get('/api/v1')
-    console.log(response)
-  }
-
-  useEffect(() => {
-    fetchData()
-  }, [])
+  const alert = useSelector((state) => state.alert)
 
   return (
     <div>